Added INetworkMapper, Added bpsim, moved network_comp to network.h
Added the INetworkMapper for altering networks.

Added PercolationMapper which is an implementation of INetworkMapper

Added the bpsim percolation simulation program

Moved the network_comp functor into network.h.  If we are stuck
with C++, we might as well make use of the STL

Changed network_comp to networkptr_gt and networkptr_lt, which
dereference the Network and compare them using Network::operator<



imported from: boykin@pobox.com--code-2005/netmodeler--pob--0.1--patch-8