added weightededge and directedweightededge
* Added weightededge and directedweighted edge.  This was *only* so
that by default we don't have to spend the memory keeping the weight
of an edge, since that is almost always 1.0 (which is the default),
this was wasting memory in the usual case, and that mattered for huge
networks.

* One should *never* cast to weighted edge, just use the getWeight
method.  This is because the heirarchy is a little stange (directedweightededge
is not a subclass of weightededge, it is a subclass of directededge).

imported from: boykin@pobox.com--code-2005/netmodeler--pob--0.2--patch-16