converted PercolationMapper to use new code.
Converted PercolationMapper to use the FilteredIterator class and the Network::remove
method.  This makes the code an excellent example of using those classes.  Also, if
either probability is 1.0, we skip that one (since we are keeping all the items,
there is no need to iterate through them).

imported from: boykin@pobox.com--code-2005/netmodeler--pob--0.2--patch-25